Entity update and Datebase update fix: wrong realtionship in `AppUser` entity class feat: Add repo for role entity refactor: database migration

+++ src/main/java/io/github/ismaele77/liveminds/repository/AppUserRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
 
-public interface AppUserRepository extends JpaRepository<AppUser, Long> {
+public interface AppUserRepository extends JpaRepository<AppUser, Integer> {
     Optional<AppUser> findByUserName(String userName);
 }
